Rename show to detail book, Pr coments
