+++ wolox-books/src/app/screens/book/book-routing.module.ts
@@ -3,14 +3,14 @@ import { BookComponent } from './book.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { AuthGuard } from 'src/app/guards/auth.guard';
-import { ShowComponent } from './components/show/show.component';
+import { DetailComponent } from './components/detail/detail.component';
 
 const routes: Routes = [
   {
     path: 'books', component: BookComponent,
     children: [
       { path: '', component: ListComponent, canActivate: [AuthGuard] },
-      { path: ':id', component: ShowComponent, canActivate: [AuthGuard] }
+      { path: ':id', component: DetailComponent, canActivate: [AuthGuard] }
     ]
   },
 ];
