+++ wolox-books/src/app/screens/book/components/detail/detail.component.html
@@ -1,9 +1,9 @@
 <div class="container">
   <div>
-    <a class="back" (click)="goBack()">
+    <button class="back" (click)="goBack()" type="button">
       <img src="../../../../../assets/back-arrow.png" alt="back-arrow">
       <span class="back-text">Atr&aacute;s</span>
-    </a>
+    </button>
     <div class="card">
       <img src="../../../../../assets/show-book-cover.png" alt="book-cover" class="book-cover">
       <img src="../../../../../assets/badge.png" alt="book-cover" class="badge">
