+++ wolox-books/src/app/store/book.reducer.ts
@@ -9,9 +9,9 @@ const initialState = {
 export function bookReducer(state: Object = initialState, action: BookActions.Actions) {
   switch(action.type) {
     case BookActions.LISTED_BOOK:
-      return Object.assign({}, state, { books: [...(action as BookActions.ListedBook).payload] });
+      return {...state, ...{ books: [...(action as BookActions.ListedBook).payload] }};
     case BookActions.SHOWED_BOOK:
-      return Object.assign({}, state, { book: (action as BookActions.ListedBook).payload });
+      return {...state, ...{ book: (action as BookActions.ListedBook).payload }};
     default:
       return state;
   }
